+++ ap/aputils/eventutils.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from collections import OrderedDict
+from collections import OrderedDict, namedtuple
 from copy import copy
 
 
@@ -147,3 +147,14 @@ class EventUtils:
         t_ev_tb.setdefault(t, []).append(ev)
 
     return (ordered_ev_list, t_ev_tb)
+
+    @staticmethod
+    def time_overlap(start1, end1, start2, end2):
+      Range = namedtuple('Range', ['start', 'end'])
+      r1 = Range(start=start1, end=end1)
+      r2 = Range(start=start2, end=end2)
+      latest_start = max(r1.start, r2.start)
+      earliest_end = min(r1.end, r2.end)
+      delta = (earliest_end - latest_start).days + 1
+      overlap = max(0, delta)
+      return overlap
